pyproject.toml
[tool.poetry]
name = "semantic-search"
version = "0.1.0"
description = ""
authors = ["Waqar351 <waqar_comsat@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
sentence-transformers = "^5.1.1"
faiss-cpu = "^1.12.0"
pypdf2 = "^3.0.1"
pdfplumber = "^0.11.7"
nltk = "^3.9.1"
sqlalchemy = "^2.0.43"
pydantic = "^2.11.9"
fastapi = "^0.117.1"
uvicorn = {extras = ["standard"], version = "^0.37.0"}
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
isort = "^6.0.1"
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
